Fix updates to multiple images when using helm values file
When updating multiple images in a Helm values file, ArgoCD would only
ever update a single image. This is because the code only takes into
account the new helm values from the last image that it processed
ignoring the changes for the previous images.

This commit fixes that behavior. The `helmNewValues` var is moved
outside of the for-loop that iterates over the app images. For each
image, the tag is updated in the `helmNewValues` var.

Signed-off-by: Lyupcho Kotev <lyupcho.kotev@codira.dev>