feat: switch to yaml.v3
In order to support round-tripping of files with multiline yaml
and anchors, move to yaml.v3. This lower-level interface
preserves comments and doesn't follow aliases by default.

Support for aliases in Helm values files is fairly minimal
and I need to add tests to demonstrate that this works;
it only follows one level of alias to avoid getting stuck
in loops.

A lot of tests needed updated because it seems yaml.v3 switched to
four spaces instead of two when marshalling by default.

Signed-off-by: Brian Ewins <bewins@nerdwallet.com>