fix: always set str tag on string scalars
When yaml.v3 parses yaml, string scalars are tagged "!!str". It is
possible that we could parse a values file which sets the version to
eg 1.23 (a float), so that when we later update the value to "v2.4"
it can't be serialized. To protect against this, set the tag whenever
we create or update scalars that are explicitly intended to be strings.

Signed-off-by: Brian Ewins <bewins@nerdwallet.com>