fix: Remove incorrect yaml.LiteralStyle from helm values write-back
This commit fixes a regression introduced between v0.14.0 and v0.17.0
where helm values files would be emptied or corrupted when using git
write-back with helmvalues: target.

The issue was caused by setting `Style: yaml.LiteralStyle` on the root
MappingNode when initializing an empty helm values structure. The
LiteralStyle is intended for scalar string values with block style (|),
not for mapping nodes. This caused the YAML encoder to produce incorrect
or empty output when marshaling the structure back to YAML.

This regression affected users with multi-source applications using
separate image-name and image-tag annotations with helmvalues write-back,
resulting in their values files being deleted on update.

Fixes: Regression between v0.14.0 and v0.17.0
Testing: All existing tests pass with this change
Signed-off-by: DeZin7 <marcusandre77@icloud.com>