force app type to helm when write back target is set to helmvalues
Previously when in a multi source application that included both helm and kustomize the app type was detected as kustomize
 even when i explicitly set the writeback target to helmvalues.

This PR automatically detects helmvalues: prefix from writeback target and sets WriteBackConfig.HelmChart to true if its found.

Similarly to wbc.KustomizeBase the app type is returned as "helm" when wbc.HelmChart is true

Previous implementation of multiple sources is naive and returns the first app type argocd finds. If it would be somehow possible
 to tell image updater what source index to prefer then this change would not be needed for me atleast. But for now this method
 seems to be okay given such logic laready exists for kustomize.

Signed-off-by: Jyrno Ader <jyrno@thorgate.eu>